Layout window (if needed) before restoring its position
With auto-layout we may have views that do not yet have the proper frame size, because they have not yet been displayed.

This was causing a problem with the way we restored the window’s frame, as we use the top/left position, but windows are anchored at the bottom/left position, so if the layout system resized our content view, the window’s top position would move (new size - old size) points.